lor_to_d {MAc}	R Documentation

Log Odds Ratio to Standardized Mean Difference (d)

Description

Converts a log odds ratio reported in the primary study to a standardized mean difference (d).

Usage

lor_to_d(lor, var.lor)

Arguments

lor
Log odds ratio reported in the primary study.

var.lor
Variance of the log odds ratio.

Value

d
Standardized mean difference (d).

var_d
Variance of d.

Author(s)

AC Del Re & William T. Hoyt

Maintainer: AC Del Re acdelre@gmail.com

References

Borenstein (2009). Effect sizes for continuous data. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta analysis (pp. 279-293). New York: Russell Sage Foundation.

See Also

d_to_g, mean_to_d, mean_to_d2, t_to_d, f_to_d, p_to_d1, p_to_d2, ancova_to_d1, ancova_to_d2, tt.ancova_to_d, f.ancova_to_d, r_to_d, p.ancova_to_d1, p.ancova_to_d2, prop_to_or, prop_to_d, r_from_chi, r_from_d, r_from_d1, r_from_t

[Package MAc version 1.1.1 Index]