omni {MAc}	R Documentation

Omnibus Effect Size (Fixed and Random Effects)

Description

Computes fixed and random effects omnibus effect size for correlations.

Usage

omni(es, var, data, type="weighted", method = "random", ztor = FALSE)

Arguments

es
r or z' effect size.

var
Variance of es.

type
weighted or unweighted. Default is weighted. Use the unweighted variance method only if Q is rejected and is very large relative to the number of studies in the meta-analysis.

method
Default is random. For fixed effects, use fixed.

ztor
Default is FALSE. If TRUE, this assumes z' (Fisher's z) was used in the es argument and the analyist would like z' to be converted to r (for interpretive purposes) after analyzing in z'.

data
data.frame with above values.

Value

Fixed and random effects:

k
Number of studies in the meta-analysis.

estimate
Unstandardized regression coefficient estimate.

se
Standard error of the estimate coefficient.

z
z-value.

ci.l
Lower 95% confidence interval.

ci.u
Upper 95% confidence interval.

p
Significance level.

Q
Q-statistic (measure of homogeneity).

df.Q
Degrees of freedom for Q-statistic.

Qp
Q-statistic p-value (assesses overall homogeneity between studies).

I2
Proportion of total variation in effect size that is due to systematic differences between effect sizes rather than by chance (see Shadish & Haddock, 2009; pp. 263).

Author(s)

AC Del Re & William T. Hoyt

Maintainer: AC Del Re acdelre@gmail.com

References

Shadish & Haddock (2009). Analyzing effect sizes: Fixed-effects models. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta analysis (pp. 257-278). New York: Russell Sage Foundation.

Examples

id<-c(1:20)
n<-c(10,20,13,22,28,12,12,36,19,12,36,75,33,121,37,14,40,16,14,20)
r<-c(.68,.56,.23,.64,.49,-.04,.49,.33,.58,.18,-.11,.27,.26,.40,.49,
.51,.40,.34,.42,.16)
mod1<-c(1,2,3,4,1,2,8,7,5,3,9,7,5,4,3,2,3,5,7,1)
dat<-data.frame(id,n,r,mod1)
dat$var.r <- var_r(dat$r, dat$n) # MAc function to derive variance
dat$z <- r_to_z(dat$r)  # MAc function to convert to Fisher's z (z')
dat$var.z <- var_z(dat$n)  # MAc function for variance of z'
dat$mods2 <- factor(rep(1:2, 10))

# Example

omni(es = z, var = var.z, data = dat, type="weighted", method = "random", ztor = TRUE)

[Package MAc version 1.1.1 Index]